# backend/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_http_product.CreateProductRequest:
    properties:
      categoryId:
        type: string
      name:
        type: string
      quantity:
        type: integer
      stockThreshold:
        type: integer
    required:
    - name
    - quantity
    - stockThreshold
    type: object
  internal_http_product.UpdateProductRequest:
    properties:
      categoryId:
        type: string
      name:
        type: string
      stockThreshold:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Stock management application API.
  title: Stockify API
  version: "1.0"
paths:
...
  /product/:
    post:
      consumes:
      - application/json
      description: Create a new inventory product with name, quantity, and threshold
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: product_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create a new product
      tags:
      - Product
  /product/{id}:
    patch:
      consumes:
      - application/json
      description: Update product name, stock threshold, or category
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: product_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update a product
      tags:
      - Product
  /product/{id}/archive:
    patch:
      description: Soft-archive a product to hide it from active listings
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: product_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Archive a product
      tags:
      - Product
  /product/{id}/reactivate:
    patch:
      description: Restore an archived product back to active status
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: product_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Reactivate a product
      tags:
      - Product
...
  /product/category/{id}:
    get:
      description: Retrieve all products in a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: list of products
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get products by category
      tags:
      - Product
  /product/low-stock:
    get:
      description: Retrieve products where quantity is below stock threshold
      produces:
      - application/json
      responses:
        "200":
          description: list of low stock products
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get low stock products
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"